requirements.txt
PySide6>=6.8.0
psycopg2-binary>=2.9.9
PyJWT>=2.10.1
bcrypt>=4.1.3
python-dotenv>=1.0.1
pytest>=8.3.3
pytest-qt>=4.4.0
openpyxl>=3.1.5
